apps/ai-cortex/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
openai==1.12.0
instructor>=1.0.0
anthropic>=0.18.0
python-multipart==0.0.6
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
